Move Tide status controller code/tests into separate files.
