+++ prow/tide/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["tide.go"],
+    srcs = [
+        "status.go",
+        "tide.go",
+    ],
     importpath = "k8s.io/test-infra/prow/tide",
     visibility = ["//visibility:public"],
     deps = [
@@ -33,7 +36,10 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = ["tide_test.go"],
+    srcs = [
+        "status_test.go",
+        "tide_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//prow/config:go_default_library",
